Refactor SquareCheckout into a module with multiple files.
