+++ app/controllers/signup/payments_controller.rb
@@ -51,7 +51,7 @@ module Signup
     end
 
     def checkout
-      SquareCheckout.from_env
+      SquareCheckout::Client.from_env
     end
 
     def are_payments_enabled?
