+++ app/lib/square_checkout/square_error.rb
@@ -1,0 +1,3 @@
+module SquareCheckout
+  class SquareError < StandardError; end
+end
