+++ test/controllers/renewal/payments_controller_test.rb
@@ -33,7 +33,7 @@ module Renewal
           member_id: @member.id,
           date: Date.current)
 
-        SquareCheckout.stub :new, mock_checkout do
+        SquareCheckout::Client.stub :new, mock_checkout do
           post renewal_payments_url, params: {membership_payment_form: {amount_dollars: "12"}}
         end
 
@@ -58,7 +58,7 @@ module Renewal
         member_id: Integer,
         date: Date)
 
-      SquareCheckout.stub :new, mock_checkout do
+      SquareCheckout::Client.stub :new, mock_checkout do
         post renewal_payments_url, params: {membership_payment_form: {amount_dollars: "12"}}
       end
